Project Euler: refactor lcm problem with int values and java 8 style

+++ src/test/java/math/BaseTest.java
@@ -80,7 +80,7 @@ public class BaseTest {
 	@Test
 	@DisplayName("Find the least common multiply")
 	void lcmTest() {
-		long[] numbers = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20};
+		int[] numbers = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20};
 		System.out.println(main.lcm(numbers));
 	}
 }
